requirements.txt
SQLAlchemy~=2.0.34
fastapi~=0.113.0
pydantic~=2.9.0
loguru~=0.7.2
python-dotenv~=1.0.1
alembic~=1.13.2
asyncpg==0.29.0
starlette~=0.38.4
firebase-admin==6.5.0
python-multipart==0.0.9